requirements.txt
apache-airflow==2.6.3
apache-airflow-providers-google==10.4.0
apache-airflow-providers-sftp==4.4.0